moves fault count summing into new provider
