+++ src/pages/test-report/components/driving-fault-summary/driving-fault-summary.ts
@@ -4,9 +4,9 @@ import { Store, select } from '@ngrx/store';
 import { StoreModel } from '../../../../shared/models/store.model';
 import { getCurrentTest } from '../../../../modules/tests/tests.selector';
 import { getTestData } from '../../../../modules/tests/test-data/test-data.reducer';
-import { getDrivingFaultSummaryCount } from '../../../../modules/tests/test-data/test-data.selector';
 import { Subscription } from 'rxjs/Subscription';
 import { getTests } from '../../../../modules/tests/tests.reducer';
+import { FaultCountProvider } from '../../../../providers/fault-count/fault-count';
 
 interface DrivingFaultSummaryState {
   count$: Observable<number>;
@@ -22,7 +22,9 @@ export class DrivingFaultSummaryComponent implements OnInit {
   subscription: Subscription;
 
   constructor(
-    private store$: Store<StoreModel>) { }
+    private store$: Store<StoreModel>,
+    private faultCountProvider: FaultCountProvider,
+  ) { }
 
   ngOnInit(): void {
     this.componentState = {
@@ -30,7 +32,7 @@ export class DrivingFaultSummaryComponent implements OnInit {
         select(getTests),
         select(getCurrentTest),
         select(getTestData),
-        select(getDrivingFaultSummaryCount),
+        select(this.faultCountProvider.getDrivingFaultSummaryCount),
       ),
     };
   }
